Update to Kotlin 1.3 RC

+++ shared/ios/src/main/kotlin/io/intrepid/pickpocket/Random.kt
@@ -2,5 +2,5 @@ package io.intrepid.pickpocket
 
 import platform.posix.arc4random_uniform
 
-actual fun randomDigit(max: Int): Int = arc4random_uniform(max)
+actual fun randomDigit(max: Int): Int = arc4random_uniform(max.toUInt()).toInt()
